refactor(action-harness): split sequencer harness modules
Move the production sequencer harness wiring out of a monolithic sequencer.rs file into focused sequencer submodules. Keep sequencer/mod.rs declarative with only module docs, module declarations, and re-exports.
